pyproject.toml
[project]
name = "exp-staging-backend"
version = "0.1.0"
description = "exp-staging-backend"
readme = "README.md"
requires-python = "==3.11.13"
dependencies = [
    "asyncio>=3.4.3",
    "diffusers>=0.34.0",
    "fastapi[standard]>=0.116.1",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "transformers>=4.54.0",
    "ultralytics>=8.3.170",
]
